pyproject.toml
[project]
name = "arena-dei-poveri"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets",
    "kagglehub",
    "marimo",
    "mistralai",
    "pandas",
    "python-dotenv>=1.2.1",
    "rich",
    "huggingface-hub>=0.27",
    "wandb",
]
